requirements.txt
eth-abi
eth-brownie>=1.11.6,<2.0.0
eth-utils
toml
toolz
tqdm
rich
click